This is a working commit. It has issues.
The database create with hsql and not with derby like specified in the
jndi.properties file. Even if the code that is loading the property file
removed, it will still use hsql. The database file will be created under
$CATALINA_HOME/data.
This indicates the jndi file is not really read under tomcat. Unlike
embedded jetty, this file is never accessed.

Things to work on:
- The "java.naming.factory.initial" is set through a system property.
  This is not the best that can be done. A better approache need to be
  taken (ie. ContextListener, or someother configuration file).